# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: The Champions Club API
  version: 1.0.0
  description: "GENERATED from tools/api/endpoints.mjs — DO NOT EDIT BY HAND. Human-readable version: API_CONTRACT.md. Schemas are derived from shared/types/*.ts (TypeScript is authoritative; inline/complex shapes may appear as free-form)."
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PageMeta:
      type: object
      properties:
        page:
          type: number
        page_size:
          type: number
        total:
          type: number
        total_pages:
          type: number
      required:
        - page
        - page_size
        - total
        - total_pages
    ApiError:
      type: object
      properties:
        success:
          description: "TS: false"
        error:
          type: object
          properties:
            code:
              description: "TS: ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: {}
          required:
            - code
            - message
      required:
        - success
        - error
...
    CourtBooking:
      type: object
      properties:
        id:
          type: string
          format: uuid
        booking_number:
          type: string
        court_id:
          type: string
          format: uuid
        booking_type:
          type: string
          enum:
            - REGULAR
            - MAINTENANCE
        member_id:
          type: string
          format: uuid
          nullable: true
        guest_name:
          type: string
          nullable: true
        guest_phone:
          type: string
          nullable: true
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        list_price:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        discount_amount:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        cancelled_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - booking_number
        - court_id
        - booking_type
        - member_id
        - guest_name
        - guest_phone
        - start_at
        - end_at
        - list_price
        - discount_amount
        - cancelled_at
        - created_at
        - updated_at
...
    PriceBreakdown:
      type: object
      properties:
        court_id:
          type: string
          format: uuid
        start_at:
          type: string
          format: date-time
        end_at:
          type: string
          format: date-time
        membership_type:
          type: string
          enum:
            - GOLD
            - SILVER
            - JUNIOR
          nullable: true
        list_price:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        discount_percent:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "18.00"
        discount_amount:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        amount_due:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        tax_rate:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "18.00"
        tax_amount:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        is_free:
          type: boolean
        plays_used_today:
          type: number
          nullable: true
        plays_allowed_per_day:
          type: number
          nullable: true
      required:
        - court_id
        - start_at
        - end_at
        - membership_type
        - list_price
        - discount_percent
        - discount_amount
        - amount_due
        - tax_rate
        - tax_amount
        - is_free
        - plays_used_today
        - plays_allowed_per_day
    BookingDetail:
      allOf:
        - $ref: "#/components/schemas/CourtBooking"
        - type: object
          properties:
            status:
              type: string
              enum:
                - CONFIRMED
                - CANCELLED
                - COMPLETED
            amount_due:
              type: string
              pattern: ^\d+\.\d{2}$
              example: "1250.00"
            amount_paid:
              type: string
              pattern: ^\d+\.\d{2}$
              example: "1250.00"
            payment_status:
              type: string
              enum:
                - PENDING
                - PARTIALLY_PAID
                - PAID
                - REFUNDED
                - NOT_REQUIRED
            court_name:
              type: string
            sport_type:
              type: string
              enum:
                - TENNIS
                - CRICKET
                - PADEL
                - BADMINTON
            member_name:
              type: string
              nullable: true
            member_code:
              type: string
              nullable: true
          required:
            - status
            - amount_due
            - amount_paid
            - payment_status
            - court_name
            - sport_type
            - member_name
            - member_code
    BookingCancelResult:
      type: object
      properties:
        booking:
          $ref: "#/components/schemas/BookingDetail"
        refund_amount:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        refund_payment_id:
          type: string
          format: uuid
          nullable: true
      required:
        - booking
        - refund_amount
        - refund_payment_id
...
paths:
...
  /bookings/price:
    get:
      operationId: bookings.price
      tags:
        - Court bookings
      summary: Price preview before confirming (applies the member plan discount and shows plays used today).
      description: "Requirements: FR-COURT-012, FR-COURT-007"
      x-roles:
        - MEMBER
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 2
      x-requirements:
        - FR-COURT-012
        - FR-COURT-007
      x-tables:
        - courts
        - memberships
        - membership_plans
      security:
        - bearerAuth: []
      parameters:
        - name: court_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: start_at
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: member_id
          in: query
          required: false
          description: staff only; omit for walk-in pricing
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/PriceBreakdown"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: COURT_NOT_FOUND | MEMBER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: INVALID_SLOT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bookings:
    post:
      operationId: bookings.create
      tags:
        - Court bookings
      summary: Book a court (member self-service, or staff for a member / walk-in / phone caller).
      description: |-
        Transaction: lock member (advisory) -> check plays-per-day (REGULAR bookings that stand, IST day, max = plan.max_plays_per_day, default 2) -> compute price -> INSERT (DB exclusion constraint `court_bookings_no_overlap` is the final arbiter; map SQLSTATE 23P01 to BOOKING_CONFLICT) -> optional payment. list_price and discount_amount are stored (snapshot); the amount due and payment_status are derived (view court_booking_totals). Expired/no membership => walk-in price. Free (amount due 0) => payment_status NOT_REQUIRED.

        Requirements: FR-COURT-004, FR-COURT-005, FR-COURT-006, FR-COURT-007, FR-COURT-008, FR-COURT-009, FR-COURT-003
      x-roles:
        - MEMBER
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 2
      x-requirements:
        - FR-COURT-004
        - FR-COURT-005
        - FR-COURT-006
        - FR-COURT-007
        - FR-COURT-008
        - FR-COURT-009
        - FR-COURT-003
      x-tables:
        - court_bookings
        - payments
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                court_id:
                  type: string
                  format: uuid
                start_at:
                  type: string
                  format: date-time
                  description: on a :00/:30 boundary, in the future, session must end by club_close_time
                member_id:
                  type: string
                  format: uuid
                  description: "Staff only: book/order on behalf of this member. MEMBER callers: omit (implied = self)."
                guest_name:
                  type: string
                  description: Walk-in name. Staff only. Required when no member_id.
                guest_phone:
                  type: string
                  description: Walk-in phone (E.164 or 10-digit). Staff only.
                payment_method:
                  type: string
                  enum:
                    - CASH
                    - CARD
                    - UPI
                    - ONLINE
                  description: omit = pay later at the desk (payment_status PENDING). MEMBER callers pay ONLINE or later at desk.
              required:
                - court_id
                - start_at
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/BookingDetail"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "402":
          description: PAYMENT_FAILED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: COURT_NOT_FOUND | MEMBER_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: COURT_UNAVAILABLE | BOOKING_CONFLICT | DAILY_BOOKING_LIMIT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: INVALID_SLOT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    get:
      operationId: bookings.list
      tags:
        - Court bookings
      summary: List bookings (day view for the desk, own bookings for a member).
      description: "Requirements: FR-COURT-011, FR-COURT-016"
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 2
      x-requirements:
        - FR-COURT-011
        - FR-COURT-016
      x-tables:
        - court_bookings
        - courts
        - members
      security:
        - bearerAuth: []
      parameters:
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: court_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: member_id
          in: query
          required: false
          description: staff only
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - CONFIRMED
              - CANCELLED
              - COMPLETED
        - name: booking_type
          in: query
          required: false
          schema:
            type: string
            enum:
              - REGULAR
              - MAINTENANCE
        - name: upcoming
          in: query
          required: false
          description: true = start_at >= now
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/BookingDetail"
                  message:
                    type: string
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bookings/{id}:
    get:
      operationId: bookings.get
      tags:
        - Court bookings
      summary: Booking detail.
      description: "Requirements: FR-COURT-011, FR-COURT-016"
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 2
      x-requirements:
        - FR-COURT-011
        - FR-COURT-016
      x-tables:
        - court_bookings
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/BookingDetail"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: BOOKING_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bookings/{id}/cancel:
    post:
      operationId: bookings.cancel
      tags:
        - Court bookings
      summary: Cancel a booking before it starts; frees the slot and refunds per policy.
      description: |-
        Only a booking that stands and starts in the future. Sets cancelled_at (derived status CANCELLED). Member callers: own bookings only, and policy decides the refund. A refund is recorded on the payment (refunded_amount, refunded_at); the booking then reports payment_status REFUNDED.

        Requirements: FR-COURT-010
      x-roles:
        - MEMBER:own
        - FRONT_DESK
        - OWNER_ADMIN
      x-owner: Dev 2
      x-requirements:
        - FR-COURT-010
      x-tables:
        - court_bookings
        - payments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refund:
                  type: boolean
                  description: "staff only: override policy (default: refund if start_at - now >= cancellation_cutoff_hours)"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/BookingCancelResult"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: FORBIDDEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: BOOKING_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: BOOKING_NOT_CANCELLABLE | INVALID_STATUS_TRANSITION
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...